list/search tree interface impl
