+++ src/com/Utility/DataStructure/ListADT.java
@@ -1,0 +1,4 @@
+package com.Utility.DataStructure;
+
+public interface ListADT<T> {
+}
