+++ src/com/Utility/DataStructure/SearchTreeADT.java
@@ -1,0 +1,9 @@
+package com.Utility.DataStructure;
+
+import com.Exception.DuplicateKeyException;
+
+public interface SearchTreeADT<K,V> {
+
+  void insert(K key, V value) throws DuplicateKeyException;
+
+}
